refactor: querygen - split out dedup code
Split out the `GenDedupEval` function in querygen to its
own file. Its a part of a move the separate generic query
generation code from user-journey-specific query data.
